requirements.txt
accelerate==1.0.1
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.5.2
astor==0.8.1
attrdict==2.0.1
babel==2.18.0
bce-python-sdk==0.9.63
beautifulsoup4==4.14.3
blinker==1.8.2
cachetools==5.5.2
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
clip-anytorch==2.6.0
colorlog==6.10.1
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.8.0.post1
filelock==3.16.1
fire==0.7.1
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.35.1
imagesize==1.5.0
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lap==0.5.12
lazy_loader==0.4
lmdb==1.8.1
lxml==6.0.2
MarkupSafe==2.1.5
matplotlib==3.7.5
modelscope==1.31.0
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
ninja==1.13.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
openvino==2024.4.0
openvino-dev==2024.4.0
openvino-telemetry==2025.2.0
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.6.1.3
paddlepaddle==2.6.2
paddlex==3.4.2
pandas==2.0.3
pdf2docx==0.5.8
pillow==10.4.0
pkg_resources==0.0.0
polars==1.8.2
premailer==3.10.0
prettytable==3.11.0
protobuf==5.29.6
psutil==7.2.1
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.20.2
pyparsing==3.1.4
pypdfium2==5.6.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
rapidfuzz==3.9.7
rarfile==4.2
regex==2024.11.6
requests==2.32.4
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tiktoken==0.7.0
timm==1.0.24
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
transformers-stream-generator==0.0.5
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.3
ujson==5.10.0
ultralytics==8.4.5
ultralytics-thop==2.0.18
urllib3==2.2.3
visualdl==2.5.3
wcwidth==0.2.14
Werkzeug==3.0.6
zipp==3.20.2
